adapt to web-sys 0.3.88 with and without unstable APIs (#59)
* adapt to web-sys 0.3.88 with and without unstable APIs

- Handle scroll_top/scroll_left returning f64 (unstable) vs i32 (stable)
- Handle client_x/client_y returning f64 (unstable) vs i32 (stable)
- Handle set_scroll_top taking f64 (unstable) vs i32 (stable)
- Adapt use_geolocation for both GeolocationPosition (unstable) and Position (stable)
- Add GeolocationCoordinates/Position/Error features to Cargo.toml
- Add check-cfg for web_sys_unstable_apis

* chore: cargo fmt

+++ crates/yew-hooks/src/hooks/use_infinite_scroll.rs
@@ -69,7 +69,10 @@ where
 
     use_event(node, "scroll", move |e: Event| {
         let element: Element = e.target_unchecked_into();
-        if element.scroll_height() - element.scroll_top() <= element.client_height() + 100 {
+        #[allow(clippy::unnecessary_cast)]
+        if (element.scroll_height() as f64) - (element.scroll_top() as f64)
+            <= (element.client_height() as f64) + 100.0
+        {
             load_more.set(true);
             debounce.run();
         }
